Move RHash from nuget package to project
