+++ NuGet.Config
@@ -3,7 +3,6 @@
 <configuration>
     <packageSources>
         <add key="nuget.org" value="https://api.nuget.org/v3/index.json" protocolVersion="3" />
-        <add key="LocalSource" value="./LocalPackages" />
         <add key="github.com/Mik1ll" value="https://nuget.pkg.github.com/Mik1ll/index.json" />
     </packageSources>
 </configuration>
