--- RHashWrapper/RHashWrapper.csproj
@@ -1,19 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net8.0</TargetFramework>
-        <LangVersion>12</LangVersion>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-        <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-        <PackageOutputPath>../LocalPackages</PackageOutputPath>
-        <Version>1.1.0</Version>
-        <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-        <!-- https://learn.microsoft.com/en-us/nuget/create-packages/native-files-in-net-packages#example-1 -->
-        <!-- <BuildOutputTargetFolder>ref;runtimes/any/lib</BuildOutputTargetFolder> -->
-    </PropertyGroup>
-
-    <ItemGroup>
-        <None Include="runtimes/**" Pack="true" PackagePath="runtimes" />
-    </ItemGroup>
-</Project>
