+++ Shizou.Server/Commands/HashCommand.cs
@@ -4,12 +4,12 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
-using RHashWrapper;
 using Shizou.Data.CommandInputArgs;
 using Shizou.Data.Database;
 using Shizou.Data.Enums;
 using Shizou.Data.Models;
 using Shizou.Server.Extensions.Query;
+using Shizou.Server.RHash;
 using Shizou.Server.Services;
 
 namespace Shizou.Server.Commands;
