+++ Shizou.Server/Shizou.Server.csproj
@@ -11,11 +11,23 @@
         <Nullable>enable</Nullable>
         <LangVersion>12</LangVersion>
         <OutputType>Library</OutputType>
+        <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+        <RuntimeIdentifiers>win-x64;linux-x64;linux-arm64</RuntimeIdentifiers>
     </PropertyGroup>
 
     <ItemGroup>
+        <None Condition="'$(RuntimeIdentifier)' == ''" Include="runtimes/**/*">
+            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+        </None>
+        <None Condition="'$(RuntimeIdentifier)' != ''" Include="runtimes/$(RuntimeIdentifier)/native/*">
+            <Link>%(Filename)%(Extension)</Link>
+            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+        </None>
+    </ItemGroup>
+
+    <ItemGroup>
         <PackageReference Include="FuzzySharp" Version="2.0.2" />
-        <PackageReference Include="JsonSchema.Net.Generation" Version="4.5.0"/>
+        <PackageReference Include="JsonSchema.Net.Generation" Version="4.5.0" />
         <PackageReference Include="Microsoft.Extensions.Hosting.Systemd" Version="8.0.0" />
         <PackageReference Include="Microsoft.Extensions.Hosting.WindowsServices" Version="8.0.0" />
         <PackageReference Include="Microsoft.VisualStudio.Threading.Analyzers" Version="17.10.48">
@@ -33,8 +45,6 @@
     </ItemGroup>
 
     <ItemGroup>
-        <PackageReference Include="RHashWrapper" Version="1.1.0" />
-        <ProjectReference Include="..\RHashWrapper\RHashWrapper.csproj" ReferenceOutputAssembly="false"/>
         <ProjectReference Include="..\Shizou.Data\Shizou.Data.csproj" />
     </ItemGroup>
