+++ Shizou.Server/runtimes/win-x64/native/librhash.lib
