+++ Shizou.sln
@@ -11,8 +11,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shizou.Blazor", "Shizou.Bla
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shizou.Tests", "Shizou.Tests\Shizou.Tests.csproj", "{55E2CB60-505D-4639-9620-39767AF09E15}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RHashWrapper", "RHashWrapper\RHashWrapper.csproj", "{D1DB5EB5-6C5C-4B46-9D79-D9B51C459023}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shizou.WebApi", "Shizou.WebApi\Shizou.WebApi.csproj", "{9FED2506-EB6E-4B95-A965-7A9529990701}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shizou.JellyfinPlugin", "Shizou.JellyfinPlugin\Shizou.JellyfinPlugin.csproj", "{4F2E9D56-ADDF-4986-B616-23F29FE83450}"
@@ -41,10 +39,6 @@ Global
 		{55E2CB60-505D-4639-9620-39767AF09E15}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{55E2CB60-505D-4639-9620-39767AF09E15}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{55E2CB60-505D-4639-9620-39767AF09E15}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D1DB5EB5-6C5C-4B46-9D79-D9B51C459023}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D1DB5EB5-6C5C-4B46-9D79-D9B51C459023}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D1DB5EB5-6C5C-4B46-9D79-D9B51C459023}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D1DB5EB5-6C5C-4B46-9D79-D9B51C459023}.Release|Any CPU.Build.0 = Release|Any CPU
 		{9FED2506-EB6E-4B95-A965-7A9529990701}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{9FED2506-EB6E-4B95-A965-7A9529990701}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9FED2506-EB6E-4B95-A965-7A9529990701}.Release|Any CPU.ActiveCfg = Release|Any CPU
